Rename common/utils.ts to common/pool.ts
