+++ server/Server.ts
@@ -16,7 +16,7 @@ if (process.env.NODE_ENV !== 'production') {
   require('dotenv').config();
 }
 
-import { pgPool } from 'common/utils';
+import { pgPool } from 'common/pool';
 const REGEN_HOSTNAME_PATTERN = /regen\.network$/;
 const WEBSITE_PREVIEW_HOSTNAME_PATTERN = /deploy-preview-\d+--regen-website\.netlify\.app$/;
 const REGISTRY_PREVIEW_HOSTNAME_PATTERN = /deploy-preview-\d+--regen-registry\.netlify\.app$/;
